+++ test-common/src/chain_env.rs
@@ -286,6 +286,7 @@ impl ChainEnv {
             storage,
             account_storage,
             None,
+            None,
         )
     }